Fixing errors of minimizing space usage code.

+++ utils/removeImageFromDB.js
@@ -3,9 +3,17 @@ const {join} = require("node:path");
 
 const removeImageFromDB = (imageName)=>
 {
+    if(!fs.existsSync(join(__dirname, "..", "uploads", imageName)))
+    {
+        return false;
+    }
+
     fs.unlink(join(__dirname, "..", "uploads", imageName), (err)=>
     {
-        if(err) throw err;
+        if(err)
+        {
+            console.log("Remove image from DB failed: ",err.message);
+        }
     })
 }
